updated all signup requirements and did separation of concerns
